feat: Add initial version of authorization
Add the initial authorization version of Umed backend. This is done by
using a decorator with a "minimum scope" as part of the parameter. This
scope can have the role and roleId required to access the resource.

To authorize, the passing token needs to have both the role and roleId.
This is done as part of initial resource creation, demonstrated in
create_patient method.

Note that this authorization method is bare-minimum and is only good for
the current small scale.

https://github.com/umed-jp/kitsune-backend/issues/3