refactor: make ResourceClient testable
This commit refactor the ResourceClient class in fhir_store so that it
can be more testable. It's done by separating the construction of
session and url to their own methods, to allow constructing test
dependency.

Also do some refactor so that fhir_configuration is no longer needed
(only url is required).

https://dev.azure.com/UMedInc/Kitsune/_workitems/edit/112