chore: migrate slot generation logic to backend
Migrate the slot logic from frontend to backend, so that we can change
the logic freely from backend side without an app update.

Note that there is a slight difference with the normal frontend logic,
in which there is no selected date time, and users will just input the
start time and end time instead.

Also, this migration will use solely "status=free" endpoint, and it will
not use as a search for FHIR.

https://dev.azure.com/UMedInc/Engineering/_sprints/taskboard/Engineering%20Team/Engineering/Sprint%2024?workitem=771