feat: prevent frontend from book unavailble slots
This fixes AB#771.

After testing with the front end given the new availble slots logic,
it's concluded that it's error-prone for the frontend to determine the
start time as the time for checking minimum booking logic, and we should
have a better approach.

The solution is to make kitsune-backend in charge of the minimum booking
time as well, meaning the backend will return only slots after booking
delay, regardless of the start time and end time from the frontend. This
basically micics the "real" slot service, in which the slot service will
automatically make slots before the delay unavailable.

Because of the change in logic (using datetime.now()), the integration
test will also need to be refactor.